+++ docs/pages/components/colorpicker/examples/ExRepresentation.vue
@@ -1,0 +1,19 @@
+<template>
+    <section>
+        <b-field label="Select a color">
+            <b-colorpicker :value="selected" representation="square" inline/>
+        </b-field>
+    </section>
+</template>
+
+<script>
+import Color from '../../../../../src/utils/color'
+
+export default {
+    data() {
+        return {
+            selected: Color.parse('red')
+        }
+    }
+}
+</script>