+++ src/components/notification/Notification.vue
@@ -18,8 +18,8 @@
                     <b-icon
                         :icon="computedIcon"
                         :pack="iconPack"
+                        :size="newIconSize"
                         both
-                        size="is-large"
                         aria-hidden/>
                 </div>
                 <div class="media-content">
@@ -32,7 +32,8 @@
                 </div>
             </div>
             <b-progress
-                v-if="progressBar"
+                v-if="autoClose && progressBar"
+                class="auto-close-progress"
                 :value="remainingTime - 1"
                 :max="duration / 1000 - 1"
                 :type="type"