+++ src/components/tooltip/Tooltip.vue
@@ -1,5 +1,5 @@
 <template>
-    <span ref="tooltip" :class="rootClasses">
+    <div ref="tooltip" :class="rootClasses">
         <transition :name="newAnimation">
             <div
                 v-show="active && (isActive || always)"
@@ -23,7 +23,7 @@
             @mouseleave="close">
             <slot ref="slot" />
         </div>
-    </span>
+    </div>
 </template>
 
 <script>
@@ -45,6 +45,10 @@ export default {
         delay: {
             type: Number,
             default: () => config.defaultTooltipDelay
+        },
+        closeDelay: {
+            type: Number,
+            default: () => config.defaultTooltipCloseDelay
         },
         position: {
             type: String,
@@ -107,7 +111,8 @@ export default {
         }
     },
     watch: {
-        isActive(value) {
+        isActive() {
+            this.$emit(this.isActive ? 'open' : 'close')
             if (this.appendToBody) {
                 this.updateAppendToBody()
             }
@@ -121,7 +126,11 @@ export default {
                 // update wrapper tooltip
                 const tooltipEl = this.$data._bodyEl.children[0]
                 tooltipEl.classList.forEach((item) => tooltipEl.classList.remove(item))
-                if (this.$vnode && this.$vnode.data && this.$vnode.data.staticClass) {
+                if (
+                    this.$vnode &&
+                    this.$vnode.data &&
+                    this.$vnode.data.staticClass
+                ) {
                     tooltipEl.classList.add(this.$vnode.data.staticClass)
                 }
                 this.rootClasses.forEach((item) => {
@@ -135,17 +144,45 @@ export default {
                         tooltipEl.classList.add(item)
                     }
                 })
-                tooltipEl.style.width = `${trigger.clientWidth}px`
-                tooltipEl.style.height = `${trigger.clientHeight}px`
+
                 const rect = trigger.getBoundingClientRect()
+
                 let top = rect.top + window.scrollY
                 let left = rect.left + window.scrollX
+
+                const quaterHeight = trigger.clientHeight / 2 / 2
+
+                switch (this.position) {
+                    case 'is-top':
+                        tooltipEl.style.width = `${trigger.clientWidth}px`
+                        tooltipEl.style.height = `0px`
+                        top -= trigger.clientHeight - quaterHeight
+                        break
+                    case 'is-bottom':
+                        tooltipEl.style.width = `${trigger.clientWidth}px`
+                        tooltipEl.style.height = `0px`
+                        top += quaterHeight
+                        break
+                    case 'is-left':
+                        tooltipEl.style.width = `0px`
+                        tooltipEl.style.height = `${trigger.clientHeight}px`
+                        break
+                    case 'is-right':
+                        tooltipEl.style.width = `0px`
+                        tooltipEl.style.height = `${trigger.clientHeight}px`
+                        left += trigger.clientWidth
+                        break
+                }
+
                 const wrapper = this.$data._bodyEl
                 wrapper.style.position = 'absolute'
                 wrapper.style.top = `${top}px`
                 wrapper.style.left = `${left}px`
+                wrapper.style.width = `0px`
                 wrapper.style.zIndex = this.isActive || this.always ? '99' : '-1'
-                this.triggerStyle = { zIndex: this.isActive || this.always ? '100' : undefined }
+                this.triggerStyle = {
+                    zIndex: this.isActive || this.always ? '100' : undefined
+                }
             }
         },
         onClick() {
@@ -181,8 +218,15 @@ export default {
         },
         close() {
             if (typeof this.autoClose === 'boolean') {
-                this.isActive = !this.autoClose
                 if (this.autoClose && this.timer) clearTimeout(this.timer)
+                if (this.closeDelay) {
+                    this.timer = setTimeout(() => {
+                        this.isActive = !this.autoClose
+                        this.timer = null
+                    }, this.closeDelay)
+                } else {
+                    this.isActive = !this.autoClose
+                }
             }
         },
         /**